Refactor AddUserName into its own directory, put calls for games into useEffect
